Second commit

+++ src/main/java/pl/sdacademy/animals/bear/BlackBear.java
@@ -3,11 +3,11 @@ package pl.sdacademy.animals.bear;
 import pl.sdacademy.clock.Clock;
 
 public class BlackBear extends Bear {
-    public BlackBear(int weight, Clock clock) {
+    public BlackBear(double weight, Clock clock) {
         super(weight, clock);
     }
 
-    public BlackBear(int weight) {
+    public BlackBear(double weight) {
         super(weight);
     }
 }
